[build] Got browserify to work
- `grunt build` creates a bundle in build/instabone.js
- `grunt browsertest` builds instabone.js and opens a test page in the browser

After a bit of playing around, I found the best approach was to expose
the index.js exports as a browser-global named Instabone. Other options
were:

- exposing `require()` and have the user do `require('instabone')`
- not exposing anything, having all browser code in a JS file

When this module is finished we'll see which approach works best. An
option is to let the user choose what kind of bundle he or she wants.